Meter total consumption auto update

+++ NETCore_MVC_Water_Company.Web/Data/SeedDb.cs
@@ -50,10 +50,10 @@
 
             if (!_context.Steps.Any())
             {
-                await AddStep(0, (float)0.3);
-                await AddStep(5, (float)0.5);
-                await AddStep(10, (float)0.8);
-                await AddStep(15, (float)1.2);
+                await AddStep(0, 0.3f);
+                await AddStep(5, 0.5f);
+                await AddStep(10, 0.8f);
+                await AddStep(15, 1.2f);
 
                 await _context.SaveChangesAsync();
             }
@@ -89,10 +89,6 @@
                 {
                     throw new InvalidOperationException($"An error occurred trying to create the default Admin Ricardo Lourenço in the seeder");
                 }
-
-                //var token = await _userHelper.GenerateEmailConfirmationTokenAsync(user);
-
-                //await _userHelper.ConfirmEmailAsync(user, token);
 
                 var isInRole = await _userHelper.IsUserInRoleAsync(user, "Admin");
